Displaying ticket name on JuiceboxRequirementCard
