+++ src/components/[guild]/Requirements/components/JuiceboxRequirementCard/JuiceboxRequirementCard.tsx
@@ -1,0 +1,46 @@
+import { HStack, Img, Skeleton, SkeletonCircle } from "@chakra-ui/react"
+import { Requirement } from "types"
+import RequirementCard from "../common/RequirementCard"
+import RequirementText from "../common/RequirementText"
+import useJuiceboxProject from "./hooks/useJuiceboxProject"
+
+type Props = {
+  requirement: Requirement
+}
+
+const JuiceboxRequirementCard = ({ requirement }: Props) => {
+  const { project, isLoading } = useJuiceboxProject(requirement?.data?.id)
+
+  return (
+    <RequirementCard requirement={requirement}>
+      <HStack spacing={4} alignItems="center">
+        <SkeletonCircle
+          minW={6}
+          boxSize={6}
+          isLoaded={!isLoading && !!project?.logoUri}
+        >
+          <Img
+            src={project?.logoUri}
+            alt={requirement.data?.id}
+            width={6}
+            borderRadius="full"
+          />
+        </SkeletonCircle>
+
+        <RequirementText>
+          {`Hold ${
+            requirement.data?.amount > 0
+              ? `at least ${requirement.data?.amount}`
+              : "any amount of"
+          } `}
+          <Skeleton display="inline" isLoaded={!isLoading}>
+            {isLoading ? "Loading..." : project.name}
+          </Skeleton>
+          {` ticket(s) in Juicebox`}
+        </RequirementText>
+      </HStack>
+    </RequirementCard>
+  )
+}
+
+export default JuiceboxRequirementCard
