+++ src/components/[guild]/Requirements/components/JuiceboxRequirementCard/index.ts
@@ -1,0 +1,3 @@
+import JuiceboxRequirementCard from "./JuiceboxRequirementCard";
+
+export default JuiceboxRequirementCard
